Make spack available on all compute instances
ParallelCluster instances are provisioned on-demand and will be set up on the
fly. As `/etc/profile.d` is not shared among compute instances, we need to add
spack setup to a shared place which will be sourced at user login. We use these
files defined on all SLURM configured ParallelCluster (default at the moment).

/etc/profile.d/slurm.csh -> /opt/slurm/etc/slurm.csh
/etc/profile.d/slurm.sh -> /opt/slurm/etc/slurm.sh

This patch addresses the issue in @mwvaughn
https://github.com/spack/spack-configs/pull/61 and makes this PR obsolete.